+++ regions/shapes/tests/test_annulus.py
@@ -90,11 +90,11 @@ class TestCircleAnnulusSkyRegion(BaseTestSkyRegion):
     wcs = make_simple_wcs(skycoord, 5 * u.arcsec, 20)
 
     expected_repr = ('<CircleAnnulusSkyRegion(center=<SkyCoord (ICRS): '
-                     '(ra, dec) in deg\n    (3., 4.)>, inner_radius=20.0 '
-                     'arcsec, outer_radius=30.0 arcsec)>')
+                     '(ra, dec) in deg\n    (3., 4.)>, inner_radius=20.0000 '
+                     'arcsec, outer_radius=30.0000 arcsec)>')
     expected_str = ('Region: CircleAnnulusSkyRegion\ncenter: <SkyCoord '
                     '(ICRS): (ra, dec) in deg\n    (3., 4.)>\ninner_radius: '
-                    '20.0 arcsec\nouter_radius: 30.0 arcsec')
+                    '20.0000 arcsec\nouter_radius: 30.0000 arcsec')
 
     def test_init(self):
         assert_quantity_allclose(self.reg.center.ra, self.skycoord.ra)
@@ -136,10 +136,10 @@ class TestEllipseAnnulusPixelRegion(BaseTestPixelRegion):
     expected_area = 7.5 * np.pi
     expected_repr = ('<EllipseAnnulusPixelRegion(center=PixCoord(x=3, y=4), '
                      'inner_width=2, outer_width=5, inner_height=5, '
-                     'outer_height=8, angle=0.0 deg)>')
+                     'outer_height=8, angle=0.0000 deg)>')
     expected_str = ('Region: EllipseAnnulusPixelRegion\ncenter: '
                     'PixCoord(x=3, y=4)\ninner_width: 2\nouter_width: 5\n'
-                    'inner_height: 5\nouter_height: 8\nangle: 0.0 deg')
+                    'inner_height: 5\nouter_height: 8\nangle: 0.0000 deg')
 
     skycoord = SkyCoord(3 * u.deg, 4 * u.deg, frame='icrs')
     wcs = make_simple_wcs(skycoord, 5 * u.arcsec, 20)
@@ -209,14 +209,14 @@ class TestEllipseAnnulusSkyRegion(BaseTestSkyRegion):
     wcs = make_simple_wcs(skycoord, 5 * u.arcsec, 20)
 
     expected_repr = ('<EllipseAnnulusSkyRegion(center=<SkyCoord (ICRS): '
-                     '(ra, dec) in deg\n    (3., 4.)>, inner_width=20.0 '
-                     'arcsec, outer_width=50.0 arcsec, inner_height=50.0 '
-                     'arcsec, outer_height=80.0 arcsec, angle=0.0 deg)>')
+                     '(ra, dec) in deg\n    (3., 4.)>, inner_width=20.0000 '
+                     'arcsec, outer_width=50.0000 arcsec, inner_height=50.0000 '
+                     'arcsec, outer_height=80.0000 arcsec, angle=0.0000 deg)>')
     expected_str = ('Region: EllipseAnnulusSkyRegion\ncenter: <SkyCoord '
                     '(ICRS): (ra, dec) in deg\n    (3., 4.)>\ninner_width: '
-                    '20.0 arcsec\nouter_width: 50.0 arcsec\n'
-                    'inner_height: 50.0 arcsec\nouter_height: '
-                    '80.0 arcsec\nangle: 0.0 deg')
+                    '20.0000 arcsec\nouter_width: 50.0000 arcsec\n'
+                    'inner_height: 50.0000 arcsec\nouter_height: '
+                    '80.0000 arcsec\nangle: 0.0000 deg')
 
     def test_init(self):
         assert_quantity_allclose(self.reg.center.ra, self.skycoord.ra)
@@ -261,10 +261,10 @@ class TestRectangleAnnulusPixelRegion(BaseTestPixelRegion):
     expected_area = 30
     expected_repr = ('<RectangleAnnulusPixelRegion(center=PixCoord(x=3, y=4), '
                      'inner_width=2, outer_width=5, inner_height=5, '
-                     'outer_height=8, angle=0.0 deg)>')
+                     'outer_height=8, angle=0.0000 deg)>')
     expected_str = ('Region: RectangleAnnulusPixelRegion\ncenter: '
                     'PixCoord(x=3, y=4)\ninner_width: 2\nouter_width: 5\n'
-                    'inner_height: 5\nouter_height: 8\nangle: 0.0 deg')
+                    'inner_height: 5\nouter_height: 8\nangle: 0.0000 deg')
 
     skycoord = SkyCoord(3 * u.deg, 4 * u.deg, frame='icrs')
     wcs = make_simple_wcs(skycoord, 5 * u.arcsec, 20)
@@ -333,14 +333,14 @@ class TestRectangleAnnulusSkyRegion(BaseTestSkyRegion):
     wcs = make_simple_wcs(skycoord, 5 * u.arcsec, 20)
 
     expected_repr = ('<RectangleAnnulusSkyRegion(center=<SkyCoord (ICRS): '
-                     '(ra, dec) in deg\n    (3., 4.)>, inner_width=20.0 '
-                     'arcsec, outer_width=50.0 arcsec, inner_height=50.0 '
-                     'arcsec, outer_height=80.0 arcsec, angle=0.0 deg)>')
+                     '(ra, dec) in deg\n    (3., 4.)>, inner_width=20.0000 '
+                     'arcsec, outer_width=50.0000 arcsec, inner_height=50.0000 '
+                     'arcsec, outer_height=80.0000 arcsec, angle=0.0000 deg)>')
     expected_str = ('Region: RectangleAnnulusSkyRegion\ncenter: <SkyCoord '
                     '(ICRS): (ra, dec) in deg\n    (3., 4.)>\ninner_width: '
-                    '20.0 arcsec\nouter_width: 50.0 arcsec\n'
-                    'inner_height: 50.0 arcsec\nouter_height: 80.0 arcsec\n'
-                    'angle: 0.0 deg')
+                    '20.0000 arcsec\nouter_width: 50.0000 arcsec\n'
+                    'inner_height: 50.0000 arcsec\nouter_height: 80.0000 arcsec\n'
+                    'angle: 0.0000 deg')
 
     def test_init(self):
         assert_quantity_allclose(self.reg.center.ra, self.skycoord.ra)